requirements.txt
flask
flask-redis
factory-boy

